rename db files appropriately consistent with their parent dir
